requirements.txt
astroid==2.0.4
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.7
isort==4.3.4
jieba3k==0.35.1
jsonrpclib-pelix==0.3.1
lazy-object-proxy==1.3.1
lxml==4.2.4
mccabe==0.6.1
newspaper3k==0.2.6
nltk==3.3
numpy==1.15.1
pika==0.12.0
Pillow==5.2.0
pylint==2.1.1
pymongo==3.7.1
python-dateutil==2.7.3
PyYAML==3.13
redis==2.10.6
requests==2.19.1
requests-file==1.4.3
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
tldextract==2.2.0
typed-ast==1.1.0
urllib3==1.23
wrapt==1.10.11
tensorflow==1.10.1
pandas==0.23.4
